delete items complete
