+++ src/App.js
@@ -9,9 +9,8 @@ import bonusItems from "./data/bonusItems";
 function App() {
   const [adoptBirdTotal, setAdoptBirdTotal] = useState(0);
   const [cart, setCart] = useState([]);
+  const [bonus, setBonus] = useState([]);
  
-
-
 
   function handleAdopt(bird) {
     if (cart.includes(bird)) {
@@ -19,18 +18,39 @@ function App() {
     }
     setCart([...cart, bird]);
     setAdoptBirdTotal(adoptBirdTotal + bird.amount);
-    // handleBonus(adoptBirdTotal, bird);
+    handleBonus(adoptBirdTotal + bird.amount)
   }
 
   function deleteBird(bird) {
     const newCart = cart.filter((birdItem) => bird.id !== birdItem.id);
     setCart(newCart);
     setAdoptBirdTotal(adoptBirdTotal - bird.amount);
-    // handleBonus(adoptBirdTotal, bird);
+    handleBonus(adoptBirdTotal - bird.amount)
   }
 
   
-  
+  function handleBonus(adoptBirdTotal) {
+
+    if (adoptBirdTotal === 0) {
+    setBonus([])
+    }    else if (adoptBirdTotal  >= 100 && adoptBirdTotal <= 300) {
+          setBonus([bonusItems[0]]);
+        } else if (adoptBirdTotal >= 300 && adoptBirdTotal <= 500) {
+          setBonus([bonusItems[0], bonusItems[1]]);
+        } else if (adoptBirdTotal >= 500 && adoptBirdTotal <= 1000) {
+          setBonus([bonusItems[0], bonusItems[1], bonusItems[2]]);
+        } else {
+          setBonus(bonusItems);
+        }
+      }
+    
+
+  function handleSubmit (e) {
+    e.preventDefault()
+    setCart([]);
+    setBonus(0)
+    alert("You have adopted birds. Thank you!")
+  }
 
  
 
@@ -43,12 +63,14 @@ function App() {
         setAdoptBirdTotal={setAdoptBirdTotal}
         deleteBird={deleteBird}
         bonusItems={bonusItems}
-        // bonus={bonus}
-        // handleBonus={handleBonus}
+        bonus={bonus}
+    
         
       />
       <BirdCard birdData={birdData} handleAdopt={handleAdopt} />
-      <Checkout />
+      <Checkout 
+      handleSubmit={handleSubmit}
+      />
     </div>
   );
 }
