+++ src/Components/Checkout.js
@@ -1,8 +1,10 @@
 
-function Checkout () {
+
+
+function Checkout ({handleSubmit}) {
     return (
-        <div>
-            <form className="form">
+        <div >
+            <form className="Checkout" >
                 <label htmlFor="firstname">First Name</label>
                 <input type="text"/>
                 <label htmlFor="lastname">Last Name</label>
@@ -11,7 +13,7 @@ function Checkout () {
                 <input type="email"/>
                 <label htmlFor="zipcode">Zip Code</label>
                 <input type="text" autoComplete="postal-code"/>
-                <button type="submit">Submit</button> 
+                <button type="submit" onClick={handleSubmit} >Submit</button> 
             </form>
         </div>
     )
